Add camera module
