+++ observation.py
@@ -1,20 +1,10 @@
-import cv2
 import time
+
+import camera
 
 START = 60
 END = 120
 INC = 20
-
-def capture_image(camera_idx):
-    camera = cv2.VideoCapture(camera_idx)
-    if not camera.isOpened():
-        print("Failed to open camera! Restart.")
-        exit()  # TODO: make exception
-    rtrn, frame = camera.read()
-    if not rtrn:
-        print("Failed to capture image!")
-    camera.release()  # Force-flush the camera
-    return frame
 
 def set_pos(pwm, pos):
     dc_range = (5, 10)
@@ -30,7 +20,7 @@ def capture_view(pwm, reverse = False) -> list:
     inc = INC if not reverse else -INC
     for pos in range(start, end, inc):
         set_pos(pwm, pos)
-        images.append(capture_image(0))
+        images.append(camera.capture_image(0))
         time.sleep(0.75)
     print(f"Captured {len(images)} images.")
     if reverse:
